Add accent debugging to spell check
Was getting some confusing errors when adding words with accents to dictionary.txt. It wouldn't run, and I wasn't sure why :) 

In other words, "ó" in the dictionary.txt file was causing the spell check to break.

Added some messaging to try to make this easier to diagnose from the Actions log.